move cache root related functions to cache.py
